// lib/types.ts
export interface Project {
  id: number;
  name: string;
  revenue: number;
  problem: string;
  solution: string;
  developer: string;
  ideation: string;
  mvp: string;
  growth1: string;
  growth2: string;
  industry: string;
  platform: string;
  productType: string;
  target: string;
  priceRange: string;
  businessModel: string;
  freePlan: string;
  pricingDetails: string;
  traffic: string;
  revenuePerTraffic: string;
  stillSolo: string;
  caseStudy: string;
}